flask-boilerplate-dashboard-argon-master/requirements.txt
flask
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy
flask_upload
gunicorn
xlrd
opencage